implement 'composer.json/composer.lock' intermediate package handling, and stability flag handling, after all
While this is no longer needed for the original use case (emitting an installation event for a package from which even those requirements can be read that were optimized away by the Composer solver because they were provided by another package, which was the case for shared extensions bundled with PHP that we needed to activate; no longer necessary now because those are 'real' virtual packages), we still need this for cases where a userland dependency declares an extension as 'provide'd. We want to attempt installation of these provided extensions after the main install has finished, but we again need to see these dependencies in the first place, and that's only possible by moving them into a package separate from the root requires.

Also necessitates mapping of stability-flags entries to requires in the (real) root requires, as stability flags are ignored anywhere except in the require section of the root file.