Use 'warned' crate for results w/ notices/warnings
The crate isn't perfect ("unwrap" name and "map_warnings" really being a "for_each"), but writing our own is for another day :)

All notices now mapped through the call chain, and in tests we now just use debug formatting for them, so we can drop strum_macros for that.

Using a Vec for notices even when in ensure_runtime_requirement where no Warned is needed (because no result value) for consistency.