Update buildpack metadata in buildpack.toml
Adjusts the buildpack `name` and `description` in `buildpack.toml`
to match the style discussed in:
https://github.com/heroku/cnb-builder-images/issues/408

These fields are used by the CNB registry and can also be seen in the
output of `pack builder inspect`. They are not used by `pack build` or
Kodon.

In addition, I've cleaned up the `[metadata]` table/subtables to match
the more concise style used by the Procfile and Python CNBs.

GUS-W-14121598.