Remove unused scripts field  (#168)
* Expose target_triple function

* Add a smoke test that passes

This test passes but does not exercise the behavior we want to test against.

* Add failing test

* Fix composer.json scripts described as objects

The previous code:

```
    #[serde_as(as = "Option<HashMap<_, OneOrMany<_, PreferOne>>>")]
    pub scripts: Option<HashMap<String, Vec<String>>>,
```

Will handle either `"scripts": {"key": "value"}` or `"scripts": {"key": ["value"]}` but not `"scripts": {"key": {"k": "v"}}`.

When json containing a map (i.e. `{}`) is parsed it fails because it is not the shape that we told serde was valid

We can fix this parse error by removing this line. By default `serde` allows serializing to a subset. i.e. if you're only expecting "name" but receive "name" and "age" it will still deserialize and quietly drop the extra information. Because the php buildpack is not using this field, we can remove it, which removes the serde error and allows tests to pass.

Close #81

* Changelog

* Apply suggestions from code review

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>
Signed-off-by: Richard Schneeman <richard.schneeman+no-recruiters@gmail.com>

* Simplify smoke test

Uses the built in libcnb `app_dir_preprocessor` to achieve a similar directory modification https://docs.rs/libcnb-test/0.28.1/libcnb_test/struct.BuildConfig.html#method.app_dir_preprocessor.

Removes the `run_shell_command` call as it's not needed, we just need to assert that the app built correctly.

---------

Signed-off-by: Richard Schneeman <richard.schneeman+no-recruiters@gmail.com>
Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>