Handle non-url "dist" in composer.lock (#176)
* Add failing test for #105

Adding a "path" type of "repositories" causes the lock parser to fail because the "url" is actually a path and not a valid URL. This reproduces one failure mode seen in 105:

```
 thread 'tests::platform::generator::make_platform_json_with_fixtures' panicked at buildpacks/php/src/tests/platform/generator.rs:37:10:
 called `Result::unwrap()` on an `Err` value: Error("relative URL without a base: \"./clock\"", line: 490, column: 9)
```

* Handle non-url "dist" in composer.lock

In #105 the error is raised because we're trying to parse "./path" into a Url type which is not valid. This field is not used in the codebase, we can remove it to prevent the error.