Include command in failure
The prior error stated "The following is the full output from the installation attempt." However, it only contained the exit status:

```rust
#[derive(Debug)]
pub(crate) enum CommandError {
    Io(io::Error),
    NonZeroExitStatus(ExitStatus),
}
```

```rust
                Err(CommandError::NonZeroExitStatus(exit_status))
```

It is modified to include the exit status and command name that was run. Since the command was already streamed, it does not duplicate that output but instead suggests the user see the stream output from earlier.