Include composer error details in error message
Previously, when an exception resulted in a non-zero exit, the error message contained no debugging information. It now specifies the command that fails, includes the exit status, and prompts the user to view the already streamed output.

The only information it's adding that wasn't there before is the exit status. However, it safeguards this code path for future modification in case someone changes the call type to timed (non-streamed), in which case it will also include the full output that was not previously visible.