Add command information to error outputs (#180)
* Show command output on composer env failure

The prior code would emit an exit status on failure. This code emits the full command and all output on failure for easier debugging.

I also updated the error message to be more specific based on what this is trying to do https://github.com/heroku/buildpacks-php/pull/179/files#diff-92cfe306b718e8b572875d34164cf6ad9b762ca2183d952e5b77a74cfdb2caa7.

* Include command in failure

The prior error stated "The following is the full output from the installation attempt." However, it only contained the exit status:

```rust
#[derive(Debug)]
pub(crate) enum CommandError {
    Io(io::Error),
    NonZeroExitStatus(ExitStatus),
}
```

```rust
                Err(CommandError::NonZeroExitStatus(exit_status))
```

It is modified to include the exit status and command name that was run. Since the command was already streamed, it does not duplicate that output but instead suggests the user see the stream output from earlier.

* Include composer error details in error message

Previously, when an exception resulted in a non-zero exit, the error message contained no debugging information. It now specifies the command that fails, includes the exit status, and prompts the user to view the already streamed output.

The only information it's adding that wasn't there before is the exit status. However, it safeguards this code path for future modification in case someone changes the call type to timed (non-streamed), in which case it will also include the full output that was not previously visible.

* Add missing commas

* Rename variant

* Refactor function

The header provided in the function is only valid for one variant, I moved it inside of that variant to reduce confusion.