Fix git@ urls in the composer.lock
Similar to the error reported in #105 and fixed in #176 this problem is caused by strictly serializing a non-url `"git@github.com:heroku/heroku-buildpack-php.git"` into a Url type which fails to parse. The field is not used and we can discard it for now. In the future we could change it to a `String` or a `serde::json::Value` if we wish to use the field. Required behavior would dictate the direction.

The test fails without this patch and passes with it.

Fix #187