Use installer plugin output FD mode for platform install progress output
The installer plugin has the ability to output the "clean" human-readable install progress to a dedicated file descriptor.

This is also how the Classic buildpack performs output of the operation.

The "raw" Composer install progress, which would be confusing to the user (since it looks similar to a userland package install), is only used for logging and possible diagnostic output on error.

This is consistent wth how the Classic buildpack produces output, and much cleaner.

GUS-W-19848805