Test for polyfill behavior
This commit https://github.com/heroku/buildpacks-php/pull/197/commits/658c0b1f7f703024147265dedc97aca2ee89a810 introduced a test for to assert that this command is not mutated and reused: https://github.com/heroku/buildpacks-php/blob/463ecfd943c1c1437d5675e4ea553d61390ff6b7/buildpacks/php/src/layers/platform.rs#L114.

This PR uses the same test (modified to use PHP 8.3 so it can run against main) to demonstrate that the test needs to be modified to protect against regression.

## Expected

If the test exercises the behavior, it will fail on main and pass on the PR

## Actual

This test passes on main and on the PR.

## Conclusion

That seems to indicate that the test does not exercise the behavior as expected. It should be modified until it fails (without rebase), then passes with rebase.

GUS-W-18505832