State the file name being over-ridden (#183)
This notice is emitted when someone provides `COMPOSER=a/different/value/here.json`. With no context, it's unclear that the default file here is `composer.json` in the root if someone is inheriting a composer project and is not familiar with that env var. While they could look it up, this provides the information inline.