Revert "Handle non-url "dist" in composer.lock (#176)"
This reverts commit a0fcff773ef8f87c8b9faff35673329c3b063021.

We do not want to have `ComposerPackageSource` with a `url` field but `ComposerPackageDist` without - the `ComposerRepository` variants have much of the same challenges regarding "non-URL" URLs, and we need the `url` fields there in various places for internal use.

Also, the same issue can occur in ComposerPackageSource, so should be fixed completely (by correctly handling "non-URL" URLs, not by simply removing the field).

The test fixture can also be massively simplified (and should not be a test case in `buildpacks/php`, but in `composer`, since that's where the issue and the code changes are located).

Also refs #105

GUS-W-18768244