Support object for 'repositories' key in composer.json
Custom repositories are typically listed in a composer.json file using an array,
see https://getcomposer.org/doc/04-schema.md#repositories.

This is the recommended way, because repositories are used in the specified
order: https://getcomposer.org/doc/articles/repository-priorities.md

There is an alternative notation for repositories, where they are listed in an
object, with a "reference name" as the key and the repository definition as the
value.

Although PHP implements "associative arrays" in a way that preserves key order,
other software that handles a composer.json may not.

While not recommended, because no ordering guarantees can be made (see note at
the end of https://getcomposer.org/doc/04-schema.md#repositories), the composer
CLI generates this notation when using 'composer config repo' for $reasons (see
https://github.com/composer/composer/issues/9918).

We have to support both notations, otherwise a user composer.json with this
notation will not parse. Internally, we only ever use the regular array form
(e.g. for our "platform.json"), and user composer.json repositories are not
processed, so the Iterator impl for ComposerRepositoriesIter delegating next()
to HashMap::values().nth() is fine (it just has to work, not be efficient) and
saves us unnecessary re-implementation.

GUS-W-18767798