Handle disabling of repositories by name if repositories is an object
In array form, repositories (think internal ones, "packagist.org") are disabled
by using an "anonymous" hash in the form {"packagist.org": false} with exactly
one key and a false value as an entry in the repositories list.

In the now-supported object form, a repository is disabled by specifying 'false'
as the value for a key (also usually only useful for "packagist.org").