Fixture based tests for composer.json and composer.lock serde handling
Most of the permutations can be placed into a single composer.json, but there are (and, in the future, likely will be more) exceptions.

For instance, older versions of Composer (before 2.8) generate an empty array instead of an empty object in the lock file if the "platform" requirements list ends up being empty (as well as some others). Our PhpAssocArray struct solves that with a serde::Deserialize impl, which of course we want to test.

GUS-W-18767784