Allow repositories as Array or Object
Previously the parser only supported composer.json with an array structure. This updates it to work with all documented types (object and array):

```
{
    "repositories": [
```

And

```
{
    "repositories": {
```

In https://getcomposer.org/doc/04-schema.md. It gives this example:

```
{
    "repositories": {
        "foo": {
            "type": "composer",
            "url": "http://packages.foo.com"
        }
    }
}
```

By inspection, the key seems to not matter i.e. the value of "foo" has no effect. Therefore this is equivalent to

```
{
    "repositories": [
        {
            "type": "composer",
            "url": "http://packages.foo.com"
        }
    ]
}
```