Re-implement disabled repository handling
Internally always converts repositories to a Vec to maintain the order at load time. This is acceptable, as an object of repositories in a composer.json can safely be re-serialized to an array.

Now really only allows a boolean false (for object notation), or a single-entry object with name as the key and boolean false as the value (for array notation), for disabling repositories.