Proof of concept: usage/help screen default entrypoint
Our buildpacks currently set `web` as the default process type, and the Procfile buildpack sets that as the default, too, if present. The process type's command wrapper (e.g. `/lifecycle/launcher/web`) is set as the `ENTRYPOINT` for the image. This wrapper sets the env vars from all buildpacks and layers, and execs the process type's command. Any arguments provided to docker run after the image name are forwarded to this command as additional arguments.

As a result, one typically cannot `docker run <image-name> bash -c "echo hello world"` without specifying `--entrypoint launcher`.

A workaround employed by some buildpacks is to define the default process type's command as `["bash", "-c"]`, and then make the default argument(s) whatever the actual command is.

However, this does not allow for additional arguments, so the above example doesn't work either (only `docker run <image-name> bash`, without further arguments).

Furthermore, having `web` be the default process type is not actually useful in practice, because a user must pass at least `-e PORT=$PORT` and `-p $LOCALPORT:$PORT` to the `docker run` command for the image to work:

    $ docker run --rm -e PORT=8080 -p 8080:8080 image-name

This proof of concept disables the default process flag for `web`, and sets a new process type, `usage`, as the default instead. This `usage` program shows a help screen with information on how to invoke the image, a list of process types, and usage information for individual process types.

A simple invocation of the image shows usage details and a list of process types:

    $ docker run --rm image-name

The `help` sub-command gives details on how to use a particular process type:

    $ docker run --rm image-name help web

In addition, unless the first argument is `help` or `--help`, the `usage` program `exec`s all remaining arguments, meaning it's possible to launch arbitrary commands out of the box:

    $ docker run --rm image-name bash -c "echo 'hello world'"
    $ docker run --rm image-name uname -a
    $ docker run --rm image-name -- help # in case 'help' is a binary on $PATH (the 'bash' help command is just a builtin)

This proof of concept is intended to showcase the soundness of this approach, and allow others to test the usability of the idea. For the final implementation, the following should probably be changed:

- move implementation to its own buildpack
- make that buildpack last in the list for all groups in the heroku builder (so that it overrides default setting of `web` by the Procfile buildpack)
- probably have that buildpack emit some special default notes about `-e` and `-p` for the `web` process case
- skip execution when building on Heroku (so that no `usage` process shows up in the process table)
- possibly convert the entrypoint `usage.sh` program from Bash to Rust

1. this implementation currently only works on projects without a `Procfile` (since the Procfile buildpack runs last and sets `web` as default)
2. Heroku looks for a `web` process type to boot an app for HTTP traffic, not the default process type, so no impact there
3. Once added to the builder, a dedicated buildpack would immediately provide basic usage info to all existing buildpacks!

GUS-W-19330133