Additional usage help for a specific process type
The idea here is to allow buildpacks to provide some more info for specific process types by placing a file (named after the process type) into a particular layer (whose name is known to the usage program, which can then find the usage info text file).

In a standalone implementation of the buildpack, the default text for the `web` process could also be different, and remind users to supply `-e PORT` and `-p` options to `docker run`.