Filter/quote/truncate platform installation error messages
When the installation of platform packages fails due to e.g. unresolvable package requirements, we show the message produced by Composer, since it contains information on e.g. unsatisfiable package versions and other hints that might be useful to the user.

In the Classic PHP buildpack, we also filter this output a little, e.g. by prefixing everything with a "quote sign" (`> `), by stripping the `heroku-sys/` prefix from package names, by dropping some irrelevant lines from the output to make it more readable, and by truncating the message at some point to make it shorter overall.

This brings full parity with the Classic filtering, and adds tests for it.

GUS-W-20086489