Drop ext-imap requirement from smoke test fixture to avoid intermittent CI failures
PHP bundled ext-imap until 8.3. From 8.4 onwards, it became an externally built extension, so we have `php` packages up to 8.3 that produce a fake bundled `ext-imap` package in the platform repo, and a real `ext-imap` for 8.4 and 8.5.

Probably due to the real `ext-imap` having a lower version number, Composer sometimes (it's not deterministic as the SAT solver basically has to toss a coin in these situations) will pick 8.3 instead of 8.5 (giving precedence to the higher `ext-imap` version number over the higher `php` version number).

In real life, this is not an issue because users can explicitly specify a higher PHP version number, but for this test fixture, we want to check that 8.5 (or another future higher version, when it lands) is available, so for now, we are best off dropping the `ext-imap` requirement.