Ensure deterministic ordering for generated "platform.json" requirements
When translating user requirements to our internal "platform" `composer.json` we now throw all requirements into an `IndexMap`, which preserves the order of insertion, instead of a `HashMap`.

Before that, when the `HashMap` got serialized to JSON, the order in which the packages occur in the generated file would random and could not be influenced by userland `composer.json` order.

There were only very few cases where this would matter in real life, such as in the following scenario, where very open constraints for `php` and `ext-imap` could lead to different versions of PHP getting installed depending on in what order the requirements from `composer.lock` got serialized into the "platform" `composer.json`.

PHP bundled `ext-imap` until 8.3. From 8.4 onwards, it became an externally built extension, so we have `php` packages up to 8.3 that produce a fake bundled `ext-imap` package in the platform repo, and a real `ext-imap` for 8.4 and 8.5.

The bundled `ext-imap` extensions have the same version number as the PHP version they are packaged with (e.g. 8.3.25); the external `ext-imap` packages for 8.4 and 8.5 have versions like 1.0.3.

This leads to a situation where if e.g. "`*`" is specified as the version number for both `php` and `ext-imap`, the Composer SAT solver could either pick a higher `php` but lower `ext-imap` version, or vice versa.

Composer will pick 8.5 and the external `ext-imap` if the php requirement is listed first, as it will prioritize a higher version number for PHP in this case:

    {
      "require": {
        "heroku-sys/php": "*",
        "heroku-sys/ext-imap": "*"
      }
    }

If `ext-imap` is listed first, it will prioritize a higher version number for `ext-imap` first, which then constrains the possible PHP versions to 8.3 or lower, so the following will install PHP 8.3 and the bundled `ext-imap`:

    {
      "require": {
        "heroku-sys/ext-imap": "*"
        "heroku-sys/php": "*",
      }
    }

Since the `IndexMap` we now use preserves order, we now always generate these entries in the same order they occur in `composer.lock`'s `platform` and `platform-dev` objects, so the user's stated requirement order is reliably preserved, leading to a deterministic version resolution outcome.

GUS-W-20610090